--- internal/services/authorization/validate/role_definition_resource_id.go
@@ -1,24 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package validate
-
-import (
-	"fmt"
-
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/authorization/parse"
-)
-
-func ValidateRoleDefinitionResourceId(input interface{}, key string) (warnings []string, errors []error) {
-	v, ok := input.(string)
-	if !ok {
-		errors = append(errors, fmt.Errorf("expected %q to be a string", key))
-		return
-	}
-
-	if _, err := parse.ParseRoleDefinitionResourceId(v); err != nil {
-		errors = append(errors, err)
-	}
-
-	return
-}
